fix: исправлено наложение системных элементов навигации на экран приложения, путём добавления к каждому экрану виджета SafeArea
