+++ frontend/lib/views/game_view/game_page_view.dart
@@ -16,54 +16,55 @@ class _GameViewState extends State<GameView> {
   @override
   Widget build(BuildContext context) {
     var scheme = Theme.of(context).colorScheme;
-    return Scaffold(
-      backgroundColor: scheme.background,
-      body: Consumer<GameModel>(
-        builder: (context, gameModel, child) {
-          return WillPopScope(
-            onWillPop: _willPopCallback,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                MoveList(gameModel),
-                Padding(
-                  padding: const EdgeInsets.symmetric(
-                    horizontal: 25,
-                    vertical: 30,
-                  ),
-                  child: SizedBox(
-                    height: 30,
-                    child: Stack(
-                      children: [
-                        CustomIconButton(
-                          iconName:
-                              "assets/images/icons/left_big_arrow_icon.svg",
-                          color: scheme.secondary,
-                          iconSize: 32,
-                          onTap: () {
-                            context.go(RouteLocations.settingsScreen);
-                          },
-                        ),
-                        const GameStatus(),
-                      ],
+    return SafeArea(
+      child: Scaffold(
+        body: Consumer<GameModel>(
+          builder: (context, gameModel, child) {
+            return WillPopScope(
+              onWillPop: _willPopCallback,
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: [
+                  MoveList(gameModel),
+                  Padding(
+                    padding: const EdgeInsets.symmetric(
+                      horizontal: 25,
+                      vertical: 30,
+                    ),
+                    child: SizedBox(
+                      height: 30,
+                      child: Stack(
+                        children: [
+                          CustomIconButton(
+                            iconName:
+                                "assets/images/icons/left_big_arrow_icon.svg",
+                            color: scheme.secondary,
+                            iconSize: 32,
+                            onTap: () {
+                              context.go(RouteLocations.settingsScreen);
+                            },
+                          ),
+                          const GameStatus(),
+                        ],
+                      ),
                     ),
                   ),
-                ),
-                const Spacer(),
-                Padding(
-                  padding: const EdgeInsets.all(30),
-                  child: ChessBoardWidget(gameModel)),
-                const SizedBox(height: 30),
-                const Spacer(),
-                Padding(
-                  padding: const EdgeInsets.all(30),
-                  child: GameInfoAndControls(gameModel),
-                ),
-                BottomPadding(),
-              ],
-            ),
-          );
-        },
+                  const Spacer(),
+                  Padding(
+                    padding: const EdgeInsets.all(30),
+                    child: ChessBoardWidget(gameModel)),
+                  const SizedBox(height: 30),
+                  const Spacer(),
+                  Padding(
+                    padding: const EdgeInsets.all(30),
+                    child: GameInfoAndControls(gameModel),
+                  ),
+                  BottomPadding(),
+                ],
+              ),
+            );
+          },
+        ),
       ),
     );
   }
