+++ frontend/lib/views/guide_view/components/components.dart
@@ -3,3 +3,4 @@ export "guide_chose_piece_button.dart";
 export "bottom_bar_guide.dart";
 export "hint_description.dart";
 export "points_indicator.dart";
+export "guide_piece_carousel.dart";
